Change account setup special folders feature to theme 2
