+++ feature/account/setup/src/debug/kotlin/app/k9mail/feature/account/setup/ui/specialfolders/SpecialFoldersFormContentPreview.kt
@@ -1,0 +1,16 @@
+package app.k9mail.feature.account.setup.ui.specialfolders
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
+
+@Composable
+@Preview(showBackground = true)
+internal fun SpecialFoldersFormContentPreview() {
+    PreviewWithTheme {
+        SpecialFoldersFormContent(
+            state = SpecialFoldersContract.FormState(),
+            onEvent = {},
+        )
+    }
+}
