+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/specialfolders/SpecialFoldersScreen.kt
@@ -4,16 +4,13 @@ import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
-import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
 import app.k9mail.core.ui.compose.common.mvi.observe
 import app.k9mail.core.ui.compose.designsystem.template.Scaffold
-import app.k9mail.core.ui.compose.theme.K9Theme
 import app.k9mail.feature.account.common.ui.WizardNavigationBar
 import app.k9mail.feature.account.common.ui.WizardNavigationBarState
 import app.k9mail.feature.account.setup.ui.specialfolders.SpecialFoldersContract.Effect
 import app.k9mail.feature.account.setup.ui.specialfolders.SpecialFoldersContract.Event
 import app.k9mail.feature.account.setup.ui.specialfolders.SpecialFoldersContract.ViewModel
-import app.k9mail.feature.account.setup.ui.specialfolders.fake.FakeSpecialFoldersViewModel
 
 @Composable
 fun SpecialFoldersScreen(
@@ -53,18 +50,6 @@ fun SpecialFoldersScreen(
             state = state.value,
             onEvent = { dispatch(it) },
             contentPadding = innerPadding,
-        )
-    }
-}
-
-@Composable
-@PreviewDevices
-internal fun SpecialFoldersScreenK9Preview() {
-    K9Theme {
-        SpecialFoldersScreen(
-            onNext = {},
-            onBack = {},
-            viewModel = FakeSpecialFoldersViewModel(),
         )
     }
 }
